% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shiny_in_rmd.R
\name{run_shiny_in_rmd}
\alias{run_shiny_in_rmd}
\title{FUNCTION_TITLE}
\usage{
run_shiny_in_rmd()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[rmarkdown]{render}}
}
